Fix disabled merge button on conference failure.
The notification for the conference failure is not notified on the
foreground call. As a result in a scenario of foreground single
party call and background conference call merge, the notifications
do not go out on the foreground call. This is because of internal
logic of merge peer and merge host that are maintained during
conference scenarios that makes the host as conference call.

Change-Id: Ia910c1b00e3b4a232153c5b3873a4ec708830f66
CRs-Fixed: 931958
Signed-off-by: Alex Naidis <alex.naidis@linux.com>