SubscriptionController: properly detect when MSIM has 1 SIM
We cannot use sSlotIdxToSubId to determine when to turn the SIM dialog
off, because when subscription information gets reloaded (on data change
for instance), we may add one, then disable SMS prompt, then add the
other, which is incorrect. Query the total number of SIMs from the info
updater.

Ticket: CYNGNOS-2185, CYNGNOS-2570

Change-Id: Ie5d9a67aed853fe87b229b9d05e74ae2393d4c38
Signed-off-by: Roman Birg <roman@cyngn.com>